:recycle: [repositories] Move fixtures `fakefetcher` and `fetchercalls` to `tests.fixtures`
